Backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3